Fix issue #141
