+++ src/NCalc/NCalc.csproj
@@ -25,6 +25,7 @@
 
 	<ItemGroup Condition="'$(TargetFramework)' == 'net462'">
 		<PackageReference Include="System.ValueTuple" Version="4.5.0" />
+		<Reference Include="netstandard" />
 	</ItemGroup>
 
 	<ItemGroup>
